+++ src/grandorgue/GOFrame.cpp
@@ -26,6 +26,7 @@
 #include "archive/GOArchiveManager.h"
 #include "combinations/GOSetter.h"
 #include "config/GOConfig.h"
+#include "dialogs/GONewReleaseDialog.h"
 #include "dialogs/GOProgressDialog.h"
 #include "dialogs/GOSelectOrganDialog.h"
 #include "dialogs/GOSplash.h"
@@ -52,6 +53,7 @@
 #include "GOPath.h"
 #include "GOProperties.h"
 #include "Images.h"
+#include "go_defs.h"
 #include "go_ids.h"
 #include "go_limits.h"
 
@@ -113,7 +115,10 @@ EVT_SLIDER(ID_METER_FRAME_SPIN, GOFrame::OnChangeSetter)
 EVT_TEXT(ID_METER_AUDIO_SPIN, GOFrame::OnSettingsVolume)
 EVT_TEXT_ENTER(ID_METER_AUDIO_SPIN, GOFrame::OnSettingsVolume)
 EVT_COMMAND(ID_METER_AUDIO_SPIN, wxEVT_SETVALUE, GOFrame::OnChangeVolume)
-
+EVT_MENU(ID_CHECK_FOR_UPDATES, GOFrame::OnUpdateCheckingRequested)
+EVT_MENU(ID_SHOW_RELEASE_CHANGELOG, GOFrame::OnNewReleaseInfoRequested)
+EVT_MENU(ID_DOWNLOAD_NEW_RELEASE, GOFrame::OnNewReleaseDownload)
+EVT_UPDATE_CHECKING_COMPLETION(GOFrame::OnUpdateCheckingCompletion)
 EVT_UPDATE_UI_RANGE(ID_FILE_RELOAD, ID_AUDIO_MEMSET, GOFrame::OnUpdateLoaded)
 EVT_UPDATE_UI_RANGE(ID_PRESET_0, ID_PRESET_LAST, GOFrame::OnUpdateLoaded)
 END_EVENT_TABLE()
@@ -251,6 +256,11 @@ GOFrame::GOFrame(
   wxMenu *help_menu = new wxMenu;
   help_menu->Append(wxID_HELP, _("&Help\tF1"), wxEmptyString, wxITEM_NORMAL);
   help_menu->Append(wxID_ABOUT, _("&About"), wxEmptyString, wxITEM_NORMAL);
+  help_menu->Append(
+    ID_CHECK_FOR_UPDATES,
+    _("&Check for updates"),
+    wxEmptyString,
+    wxITEM_NORMAL);
   m_panel_menu = new wxMenu();
 
   wxMenuBar *menu_bar = new wxMenuBar;
@@ -496,6 +506,12 @@ void GOFrame::UpdateVolumeControlWithSettings() {
 }
 
 void GOFrame::Init(const wxString &filename, bool isGuiOnly) {
+  if (m_config.CheckForUpdatesAtStartup()) {
+    // Start update checker thread that will fire events to this frame
+    m_UpdateCheckerThread = GOUpdateChecker::StartThread(
+      GOUpdateChecker::CheckReason::STARTUP, this);
+  }
+
   m_IsGuiOnly = isGuiOnly;
   Show(true);
 
@@ -1056,6 +1072,11 @@ void GOFrame::OnMenuClose(wxCommandEvent &event) {
 }
 
 bool GOFrame::CloseProgram(bool isForce) {
+  if (m_UpdateCheckerThread) {
+    // Stop update checking so that we can safely destruct the thread
+    m_UpdateCheckerThread->Stop();
+  }
+
   bool isClosed = CloseOrgan(isForce);
 
   if (isClosed)
@@ -1351,6 +1372,70 @@ void GOFrame::OnRenameFile(wxRenameFileEvent &event) {
     wxRemoveFile(filepath.GetFullPath());
 
   GOSyncDirectory(filepath.GetPath());
+}
+
+void GOFrame::OnUpdateCheckingRequested(wxCommandEvent &event) {
+  if (m_UpdateCheckerThread) {
+    // Update checking is already in progress, ignore request
+    return;
+  }
+  wxBeginBusyCursor();
+  m_UpdateCheckerThread = GOUpdateChecker::StartThread(
+    GOUpdateChecker::CheckReason::USER_REQUEST, this);
+}
+
+void GOFrame::OnUpdateCheckingCompletion(
+  GOUpdateChecker::CompletionEvent &event) {
+  // Remove update checker thread
+  m_UpdateCheckerThread->Wait();
+  m_UpdateCheckerThread = nullptr;
+
+  // Handle result
+  const GOUpdateChecker::Result &result = event.GetResult();
+  if (result.checkReason == GOUpdateChecker::CheckReason::USER_REQUEST) {
+    wxEndBusyCursor();
+    if (result.successful) {
+      if (result.updateAvailable) {
+        GONewReleaseDialog dialog(
+          this, m_config, event.GetResult().latestRelease);
+        if (dialog.ShowModal() == wxID_OK) {
+          m_config.Flush();
+        }
+      } else {
+        wxMessageBox(
+          _("You are using the latest GrandOrgue version"),
+          _("Update checker"));
+      }
+    } else {
+      wxMessageBox(
+        _("Failed to check for updates: %s").Format(result.errorMessage));
+    }
+  } else if (
+    result.checkReason == GOUpdateChecker::CheckReason::STARTUP
+    && result.successful && result.updateAvailable) {
+    m_StartupUpdateCheckerResult = result;
+    wxMenu *updateAvailableMenu = new wxMenu();
+    wxString versionChangeStr
+      = wxString::Format("%s -> %s", APP_VERSION, result.latestRelease.version);
+    updateAvailableMenu->Append(ID_SHOW_RELEASE_CHANGELOG, versionChangeStr);
+    updateAvailableMenu->Append(
+      ID_SHOW_RELEASE_CHANGELOG, _("&Show changelog"));
+    updateAvailableMenu->Append(ID_DOWNLOAD_NEW_RELEASE, _("&Download"));
+    wxMenuBar *menuBar = GetMenuBar();
+    menuBar->Append(updateAvailableMenu, _("&Update available!"));
+  }
+}
+
+void GOFrame::OnNewReleaseInfoRequested(wxCommandEvent &event) {
+  GONewReleaseDialog dialog(
+    this, m_config, m_StartupUpdateCheckerResult.latestRelease);
+  if (dialog.ShowModal() == wxID_OK) {
+    m_config.Flush();
+  }
+}
+
+void GOFrame::OnNewReleaseDownload(wxCommandEvent &event) {
+  GOUpdateChecker::OpenDownloadPageInBrowser();
 }
 
 bool GOFrame::InstallOrganPackage(wxString name) {