+++ src/grandorgue/GOFrame.h
@@ -19,6 +19,7 @@
 #include "modification/GOModificationListener.h"
 #include "size/GOResizable.h"
 #include "threading/GOMutex.h"
+#include "updater/GOUpdateChecker.h"
 
 #include "GOEvent.h"
 
@@ -70,6 +71,8 @@ private:
   bool m_MidiMonitor;
   bool m_isMeterReady;
   bool m_IsGuiOnly;
+  std::unique_ptr<GOThread> m_UpdateCheckerThread;
+  GOUpdateChecker::Result m_StartupUpdateCheckerResult;
 
   // to avoid event processing when the settings dialog is open
   bool m_InSettings;
@@ -161,6 +164,11 @@ private:
   void OnSetTitle(wxCommandEvent &event);
   void OnMsgBox(wxMsgBoxEvent &event);
   void OnRenameFile(wxRenameFileEvent &event);
+
+  void OnUpdateCheckingRequested(wxCommandEvent &event);
+  void OnUpdateCheckingCompletion(GOUpdateChecker::CompletionEvent &event);
+  void OnNewReleaseInfoRequested(wxCommandEvent &event);
+  void OnNewReleaseDownload(wxCommandEvent &event);
 
   bool CloseOrgan(bool isForce = false);
   bool CloseProgram(bool isForce = false);