+++ src/grandorgue/dialogs/settings/GOSettingsOptions.h
@@ -1,6 +1,6 @@
 /*
  * Copyright 2006 Milan Digital Audio LLC
- * Copyright 2009-2023 GrandOrgue contributors (see AUTHORS)
+ * Copyright 2009-2024 GrandOrgue contributors (see AUTHORS)
  * License GPL-2.0 or later
  * (https://www.gnu.org/licenses/old-licenses/gpl-2.0.html).
  */
@@ -73,6 +73,7 @@ private:
   wxChoice *m_Language;
   wxSpinCtrl *m_MetronomeMeasure;
   wxSpinCtrl *m_MetronomeBPM;
+  wxCheckBox *m_CheckForUpdatesAtStartup;
 
   wxString m_OldLanguageCode;
   unsigned m_OldChannels;